Split MP registration off into a separate page: MPRegistrationVerificationPage.xaml with matching view model. Compiling, but Navigation not functional.
