+++ RightToAskClient/RightToAskClient/RightToAskClient/RightToAskClient.csproj
@@ -207,6 +207,7 @@
     <None Remove="Resources\server.config" />
     <None Remove="Resources\serverconfig.json" />
     <EmbeddedResource Include="Resources\serverconfig.json" />
+    <EmbeddedResource Remove="Views\MPRegistrationPage.xaml" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Xamarin.Essentials">
