+++ RightToAskClient/RightToAskClient/RightToAskClient/Views/AccountPage.xaml
@@ -22,10 +22,9 @@
             <views:UserAccountInfoView/>
             <Label Text="{Binding ReportLabelText}" Style="{StaticResource ReportLabel}"/>
             <Button Text="{xct:Translate UpdateAccountButtonText}" Command="{Binding UpdateAccountButtonCommand}" Style="{StaticResource DoneButton}" IsVisible="{Binding ShowUpdateAccountButton}"/>
-            <ScrollView>
             <StackLayout Orientation="Vertical"> 
             <!-- TODO Unfortunately I think this is getting drawn, then made invisible, hence creating a lot of whitespace -->
-            <StackLayout Orientation="Vertical" VerticalOptions="Center" BackgroundColor="Aqua">
+            <StackLayout Orientation="Vertical" VerticalOptions="Center">
                 <StackLayout Orientation="Horizontal" HorizontalOptions="Center">
                     <Button Text="{xct:Translate CompleteAccountCreationButtonText}" Style="{StaticResource DoneButton}" Command="{Binding SaveButtonCommand}" HorizontalOptions="Center" IsVisible="{Binding ShowUpdateAccountButton, Converter={StaticResource Key=cnvInvert}}"/>
                 </StackLayout>
@@ -34,31 +33,11 @@
             <!-- <Button Text="{xct:Translate UpdateMPsButtonText}" Style="{StaticResource DoneButton}" Command="{Binding UpdateMPsButtonCommand}" IsVisible="{Binding ShowUpdateAccountButton}"/> -->
             <StackLayout Orientation="Horizontal" HorizontalOptions="CenterAndExpand">
             <Label Text="{xct:Translate MPRegistrationOptionText}" VerticalOptions="Center" />
-            <Button Text="{Binding RegisterMPButtonText}" Command="{Binding RegisterMPButtonCommand}" IsVisible="{Binding ShowRegisterMPButton}" Style="{StaticResource DoneButton}" HorizontalOptions="Center"/>
+            <Button Text="{Binding RegisterMPButtonText}" Command="{Binding ChooseMPToRegisterButtonCommand}" IsVisible="{Binding ShowRegisterMPButton}" Style="{StaticResource DoneButton}" HorizontalOptions="Center"/>
             </StackLayout>
-            <Label Text="{Binding RegisterMPReportLabelText}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
+            <!-- <Label Text="{Binding RegisterMPReportLabelText}" IsVisible="{Binding ShowRegisterMPReportLabel}" /> -->
             <Label Text="{Binding MPRepresentingName}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
-            <!-- <Label Text="{Binding Source = {Binding MPRepresentingName, StringFormat='{xct:Translate YouHaveChosen} {0}'}}" IsVisible="{Binding ShowRegisterMPReportLabel}" /> -->
-            <StackLayout Orientation="Horizontal" HorizontalOptions = "CenterAndExpand" IsVisible="{Binding ShowRegisterMPReportLabel}" >
-                <Label Text="{xct:Translate ThatsMe}"/>
-                <Switch></Switch>
-                <Label Text="{xct:Translate IAmAStaffer}"/>
             </StackLayout>
-            <Label Text="{xct:Translate EmailEntry}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
-            <StackLayout Orientation="Horizontal" IsVisible="{Binding ShowRegisterMPReportLabel}" >
-                <Entry Placeholder="{xct:Translate UserName}" Style="{StaticResource NormalEntry}" />
-                <Label Text=" @ " HorizontalOptions="Center"/>
-                <!-- <Picker x:Name="DomainPicker" ItemsSource = "models:ParliamentData.Domains"></Picker> -->
-                <Picker x:Name="DomainPicker" Title="{xct:Translate ChooseParliamentaryEmailDomain}" ></Picker> 
-            </StackLayout>
-            <Button  Command ="{Binding SendMPVerificationEmailCommand}" IsVisible="{Binding ShowRegisterMPReportLabel}" Style="{StaticResource GreenButton}" Text="{xct:Translate SendVerificationPIN}"></Button>
-            <Label Text="{xct:Translate SubmitPINButtonText}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
-            <StackLayout Orientation="Horizontal" HorizontalOptions="CenterAndExpand" IsVisible="{Binding ShowRegisterMPReportLabel}">
-                <Entry Text="{Binding MPRegistrationPIN}" Style="{StaticResource NormalEntry}" Placeholder="______" />
-                <Button Command="{Binding SubmitMPRegistrationPINCommand}" Text="{xct:Translate DoneButtonText}" Style="{StaticResource DoneButton}" />
-            </StackLayout>
-            </StackLayout>
-            </ScrollView>
         </StackLayout>
     </ContentPage.Content>
 </ContentPage>
