+++ RightToAskClient/RightToAskClient/RightToAskClient/Views/MPRegistrationVerificationPage.xaml
@@ -1,0 +1,44 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
+             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+             xmlns:xct="http://xamarin.com/schemas/2020/toolkit"
+             xmlns:views="clr-namespace:RightToAskClient.Views"
+             xmlns:vm="clr-namespace:RightToAskClient.ViewModels"
+             xmlns:local="clr-namespace:RightToAskClient"
+             xmlns:models="clr-namespace:RightToAskClient.Models;assembly=RightToAskClient"
+             x:Class="RightToAskClient.Views.MPRegistrationVerificationPage"
+             x:DataType="vm:MPRegistrationVerificationViewModel"
+             Shell.FlyoutBehavior="Flyout"
+             Title="{Binding Title}">
+    <ContentPage.BindingContext>
+        <vm:MPRegistrationVerificationViewModel />
+    </ContentPage.BindingContext>
+    <ContentPage.ToolbarItems>
+        <ToolbarItem Text="{xct:Translate InfoPopupButtonText}" Command="{Binding InfoPopupCommand}" x:DataType="vm:BaseViewModel" 
+                     IconImageSource="{local:ImageResource RightToAskClient.Images.help_75.png}"/>
+    </ContentPage.ToolbarItems>
+    <ContentPage.Content>
+        <StackLayout Margin="20" VerticalOptions="Start">
+            <Label Text="{Binding MPRepresentingName}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
+            <!-- <Label Text="{Binding Source = {Binding MPRepresentingName, StringFormat='{xct:Translate YouHaveChosen} {0}'}}" IsVisible="{Binding ShowRegisterMPReportLabel}" /> -->
+            <StackLayout Orientation="Horizontal" HorizontalOptions = "CenterAndExpand" IsVisible="{Binding ShowRegisterMPReportLabel}" >
+                <Label Text="{xct:Translate ThatsMe}"/>
+                <Switch></Switch>
+                <Label Text="{xct:Translate IAmAStaffer}"/>
+            </StackLayout>
+            <Label Text="{xct:Translate EmailEntry}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
+            <StackLayout Orientation="Horizontal" IsVisible="{Binding ShowRegisterMPReportLabel}" >
+                <Entry Placeholder="{xct:Translate UserName}" Style="{StaticResource NormalEntry}" />
+                <Label Text=" @ " HorizontalOptions="Center"/>
+                <!-- <Picker x:Name="DomainPicker" ItemsSource = "models:ParliamentData.Domains"></Picker> -->
+                <Picker x:Name="DomainPicker" Title="{xct:Translate ChooseParliamentaryEmailDomain}" ></Picker> 
+            </StackLayout>
+            <Button  Command ="{Binding SendMPVerificationEmailCommand}" IsVisible="{Binding ShowRegisterMPReportLabel}" Style="{StaticResource GreenButton}" Text="{xct:Translate SendVerificationPIN}"></Button>
+            <Label Text="{xct:Translate SubmitPINButtonText}" IsVisible="{Binding ShowRegisterMPReportLabel}" />
+            <StackLayout Orientation="Horizontal" HorizontalOptions="CenterAndExpand" IsVisible="{Binding ShowRegisterMPReportLabel}">
+                <Entry Text="{Binding MPRegistrationPIN}" Style="{StaticResource NormalEntry}" Placeholder="______" />
+                <Button Command="{Binding SubmitMPRegistrationPINCommand}" Text="{xct:Translate DoneButtonText}" Style="{StaticResource DoneButton}" />
+            </StackLayout>
+        </StackLayout>
+    </ContentPage.Content>
+</ContentPage>
